ssplit: prefix-files download, test-cases
Adds ssplit non-breaking-prefix file among "model" downloads. Uses the
said file in the common args for both File based and ByteArray based
loads for bergamot translator.

To test this ssplit prefix file, we require a file that is different
from bergamot, where non-breaking prefixes appear, and therefore a test
case using ssplit's English tests are translated through
bergamot-translator and the source-sentences in the Response information
is asserted to be consistent with expected output.